Remove reload: true from associations
This is significantly slowing down the good_nomenclature endpoints and
was leading to 100s of queries per commodity - our most used endpoint.

Essentially, by overriding the association and specifying reload: true
we're ignoring anything that was cached as a result of using the
Commodity#eager calls and costing us the earth.

I can't see how this would impact the Timemachine unless the Timemachine
plugin isn't fit for purpose/needs a wider review to work with eager
load caching (see the commits that originally put this reload in).

We saw issues with the commodities v2 controller as a result of
Bitzesty disabling fragment caching in the backend. The result of
intense queries in production weren't being cached and we are currently
creating a backlog of queries that need to be executed leading to > 60
second client timeouts.