Adds Clearable concern to all Sequel::Model classes
If a client of a model wants to clear the association cache, there is
currently no public api that enables this behaviour.

We typically want to clear it in the tests because we run the same
queries (at least as they appear to Sequel) multiple times and expect different results.

This has been specifically highlighted in the tests around the
TimeMachine which doesn't update the cache key and therefore looks like
the same query.

We can get around this by adding behaviour to Sequel::Model classes
which clears the cache after every example and by making sure that we
don't have multiple TimeMachine expectations in the same example (see
subsequent commit for this change).