Removes duplicating tests which create false negatives
Within the same example we remove a test which essentially duplicates
the same test of the TimeMachine just for a different date.

This is bad practice but the reason for the removal is because
Sequel::Associations::AssociationReflections are defined on the model
which carry a cache of each of the associations which doesn't recognise
TimeMachine changes in an eager load query. This behaviour doesn't
happen in the wild (as can be seen with the huge graph of eager
associations produced when pulling our measures/commodities and their
children in various controllers).