Clears association cache on all known problem models before a request
We know that association queries get cached and don't recognise that the
time machine should represent a change in the stored query.

Weirdly this affects only a subset of models that are invoked with the
TimeMachine.

I'd need to further dig into the specifics to understand why this is but
I know for a fact that the class level cache needs invalidating when we
want to change the actual date.

It's important to know that data gets stored in lexical scope under an
instance of a Sequel query result not in the class cache that we're
clearing here. What we're clearing is explicitly the query itself which
is reused between runs to (I suspect) save Sequel a smidge of time in
generating the output sequel as driven by invoking the sequel model and
dataset methods.