Surface measurement units for the duty calculator
We need to know all the possible measurement units

The logic will be to filter down the list of measures for the given
geographical area, then return the measures that can be expressed as
having units (expressable measures) and then present those measures in a
way that is suitable for the duty calculator.

This presentation is done via a service object which we are using to
orchestrate the filtering and presentation of the measures at the domain
layer.