Handle extra qualifier in unit code key.
In the wild we retrieve the unit from the measure units file using the
code and the qualifier concatenated. This enables us to enrich the
descriptions a bit as they relate to the specific measure
component/measure condition component.

This commit changes the key that we use to look up the measurement unit
code in fallback scenarios.

Our specific example `Factor` exists in the database but does not match
because it's code is FC1 and we were passing in the full key FC1X

```
"500 - Internal Server Error: Requested invalid measurement unit: FC1X"
```

This commit also includes some small refactor of the spec file around
linting violations