Removes cached nomenclature
HiddenGoodsNomenclature is a table containing around 10 rows normally
and doesn't need to be cached.

Adding declarable? calls to the commodity search serializer calls the
Heading#commodities_dataset method which ends up caching the
HiddenGoodsNomenclature.codes result before we've actually inserted the
HiddenGoodsNomenclature row in our HeadingsController test of
hidden_goods_nomenclature.

There are a couple of things that are true here:

1. The tests shouldn't be affected by redis caching.
2. HiddenGoodsNomenclature.codes shouldn't be cached (we're talking
   about 10 rows and a very light query)
3. The amount of meta programming in this application is
   insane/confusing and I don't believe was understood by the people
   writing the code in the first place

This commit fixes 2 and has a knock on affect of eliminating the race
condition witnessed in the v1 hidden_goods_nomenclature test