Extend list of clearable models
We have issues with the association cache (which caches the result of a
relational query on the class singleton for the model).

We need to make sure that this cache is fresh for every thread (requests
execute inside of a single puma thread) so that we avoid collisions in
which the wrong start and end dates are used for related objects.

We have a need to review the time machine plugin since it doesn't marry
well with sequel currently.

This change means that we stop the caching issue and have time for this
review