HOTT-1036: Don't reload database records in mapper
This was introduced in 20dd0bdf (PR 259) but it causes an N+1 whenever Commodity#children is requested

It was introduced as part of the Changes work _but_ I can't find
anything to suggest that its ever used as part of that work. Specs
still pass fine without the reloads and we didn't previously have
those reloads so I'm working on the assumption it was accidentally
introduced as part of the development process.