Surface resolved measure components for the duty calculator
This commit surfaces measure components that have been resolved for a
measure as part of the meursing resolution process.

I've had to simulate the behaviour of Active Record and define an id
method on the MeasureComponent model since the json api gem we're using
doesn't play well with passing a block to the relationship when building
the id and type relationship hash.

I've removed this id construction from the serializer since this is no
longer necessary and duplicates what already exists on the model.