Handle missing measurement units in the database more gracefully
Measurement units are used by the duty calculator to record inputs that
can later be used in different types of measure component to do a duty
calculation and work out the duties a trader needs to pay.

We have been seeing some components that refer to measurement units that
do not even exist in the db and these should be handled more gracefully
to avoid the backend response causing failed journeys for the user.

Even if these units do not exist in the database we can still do duty
calculations against them - they just won't be semantically enriched for
the user in the duty calculator measure amount input step.