HOTT-1084: Migrate to using callbacks for specific behaviours on specific mappers
We are in a situation where, because of CDS non-compliance with
taric data structure and processing standards we need to special
snowflake a bunch of behaviour when specific kinds of entity are
imported.

This change makes this whole process a bit neater by supplying a before
insert callback mechanism which (as the name suggests) enables us to
implement root- and branch- specific behaviour before a given entity
gets has an operation record created.

Following this specific change will be more changes to the MeasureMapper
around FootnoteAssociationMeasures