Make rake tasks list properly and fix eager loading in development
We depend on a bunch of models being loaded and haven't got explicit
requires that do this to enable doing various things with these models.

This includes things like unrestricting the primary key of each model so
the import process can update them or enumerating the models that are backed by oplogs
so we can clean oplog tables of rolled back files.

I've also updated a rake task not to try and load a constant and assign
a class variable to it before we've had a chance to load the environment