HOTT-1104 Run database migrations on own tasks app
This avoids the current 'race' scenario triggered by a newly deployed app reading the database structure on boot.

In order to perform database migrations, we need an 'app' running the latest image. _BUT_ if that app starts, it will start with the old database structure because migration has not yet happened. Best case scenario is sidekiq boots but with an old structure, worst case is we've added a new db model and the missing table prevents sidekiq from booting.

This change introduces a new app, the same size as the sidekiq worker app but without it ever starting (the `--task` option to cf push). This app can be updated, then used to migrate the database before updating the worker and web apps which will boot, but by that point migrations have happened and there should be no errors.

This 'tasks' app can also be used for running things like rake tasks if necessary.