HOTT-1286: Extend download date range logic to be 20 days ago
Now that we've removed the missing files functionality and we know that
the files can come out of order we need to try and download previously
unavailable files which now exist. Our tolerance is set to the last 20
days just as a starting point.

I've applied this behaviour to the CdsUpdate type to keep things
consistent between the implementations.