HOTT-1250: Fixes broken heading api
When the heading hasn't been preloaded, we try to serialize using the
json api fast serializer a Mash which calls Array(our_mash) on the
intermediate serialized result. We have an intermediate result because
we use a different structure for the ES cache DDL to the JSON API
schema.

We have a special snowflake Mash implementation that was nested in the
SearchClient and generically named Response. This was implemented
because of an incompatibility with the JSON API gem we're using where
Array() implicitly calls to to_a on the Hashie::Mash and creates the
wrong result for includes.

I've moved this Mash implementation, added specs for it and made sure
that we can properly serialize all Mashes we instantiate.