HOTT-1318: Fix broken measure filtering
The distinct measure regulation query needs to take account of the
regulation role in order to uniquely identify the modification or the
base regulation measures.

This is then used downstream to pull out the correct measures that are
applicable based on the TimeMachine as applied to the effective end
date.

This was subtle and unnoticed because for the most part regulation ids
are unique between the base_regulations and modification_regulations
views.