HOTT-1270 Use a connection pool for RedLock
I've observed lockups in Redlock trying to acquire a lock. The usage of Redlock is definitely not thread safe, and its being used in a threaded environment. That may not be the source of the lock ups but adding thread safety may resolve the problem and shouldn't cause other issues.