HOTT-1316: Fixes broken ExactSearch behaviour
Currently, if we have a title value that has any capitals in it we cannot
match the result with an ExactSearch because the input query gets
downcased in the constructor of BaseSearch.

The net effect of this is to prevent us from being able to query search
references until the ES index for search references is repopulated the
following morning. We see the reference in search suggestions but cannot
fetch the result.

After this change we can pull out search results without needing to
touch the ES index rows.