Stop playing whackamole with indent reloads
We currently pull out commodity indents (this does not apply to chapter/heading)
and very occasionally have the come into where these indents haven't been properly stored
against the related object - propagating errors of methods raised
against nil. This leads to intermittent failures and
we've opted to tidy up those specific places where the failure occurs.

The approach in this commit is to just globally reload by replacing the
delegating number_indents method when the nomenclature indents haven't
been loaded properly.

There should never actually be a variant where the commodity doesn't
have corresponding indents.

This is not the correct approach but is the most pragmatic for now - we
have it in our backlog to do a deep dive on why reloads are occasionally
required.