HOTT-1393 Fixed SQL when nil reduction_indicator
This solves a flaky spec (and possible real world bug)

This is fairly frequently reproducible (maybe 2 in 3 runs) at 17858ab8 using

`bundle exec rspec --seed=8725 spec/services/meursing_measure_finder_service_spec.rb`

The underlying issue is if reduction_indicator is nil, the finder produces an incorrect SQL query, using `reduction_indicator = NULL` instead of `reduction_indicator IS NULL`

This is triggered intermittently because we are using `.sample` in the `Measure` factory for `reduction_indicator` and one of the possible values is `nil`. When this value (nil) is chosen, the spec fails.