HOTT-1505 Refactor Updates Synchronize Worker
Moved the removal of Cache and triggering of reindexing from the
TariffSynchronizer up to the worker driving it.

There are 2 scenarios
1. Tariff Synchronizer is being used asynchronously via Sidekiq, in
which case this doesn't change anything but adds flexibility
2. Tariff Synchronizer is being called directly, synchronously (eg from
the rails Console) in which case the calls to schedule the cache
clearance can happen the same way.

The reason for changing this is because currently any additional steps
in the ordering of the ETL process as a whole (download/insert data +
cache invalidation + search reindexing), requires inserting every
additional step into the TariffSynchronizer. Moving this out and
integrating the sequence further up in the stack (ie the background job)
helps with separation of concerns and makes it easier to extend/change.