HOTT-1665: Adjust off-peak rules for autoscaling production
We never really scale above the minimum 8 nodes during the week unless
there is an exceptional event (e.g. someone scraping the api/calling an
expensive endpoint across a range of different resources - iterate over
all subheadings, etc).

Given this, we should be safe to change the off-peak rules such that
when the number of active users diminishes we can reduce the minimum
number of active nodes by 4.

We've also noticed that at weekends the number of users is about half
that during peak hours (06:00 to 18:00) of the week so we should be
safe to reduce the minimum number of active nodes by 2.

Over time this should save money.