HOTT-1676 Ignore index failures from Sequel ES ext
If an index is missing silently ignore the failure rather then
triggering an exception.

This is specifically to all for the scenario where we deploy new indexes
which have yet to be created, these will get created automatically
overnight as part of the sync then reindex process but the Sequel ES
extension may be triggered prior to that via the ETL Sync.