HOTT-1741: Adds dynamic and static filters to goods_nomenclature_query
This includes a change to the behviour of the must/should query construction
such that when the spacy tokeniser only returns adjectives and verbs
(the should part) we force doing a must query.

This was required to make sure that filters work under all circumstances
since without a must filters take precedence over the initial query.