HOTT-1832 Fix quota search pagination
The scope was not excluding quota definitions which had been end dated.

This was originally resolved by using compact to strip those records
which returned nil when fetched via association. That leaves the
`.count` on the scope returning an incorrect value, resulting in
incorrect pagination.

Instead I've changed the scope to exclude end dated quota definitions,
which means the `@scope.count` matches the results returned.