HOTT-2700 Fix database cleaner with rspec `around` hook
If factories are created in an RSpec `around` hook, they continue to
persist after the RSpec example completes and DatabaseCleaner has
supposedly cleaned the database.

This is because DatabaseCleaner currently runs in the `before` and
`after` RSpec hooks, which run after the `around` hook.

This means the database transaction is started after the data is created
and so the transaction rollback doesn't actually remove it.

Moving the transaction start to an `around` hook (and hence higher
priority) solves this problem.