HOTT-3118 Refactored #with_leaf_column dataset method
Previously this was doing a full table scan (albeit very quickly). This
meant it only took half a second to compute `leaf` for all commodities
but still took half a second for only one commodity.

Solving this means moving the group outside the nested query, which
means refactoring some of the code.

Now this is still quick for all commodities but only takes 3ms for a
single commodity