HOTT-3038 Amend recursive descendant populator
If all descendants skipped an indent level, then the (grand)children
would be allocated as being the children. This matched the behaviour of
the existing GoodsNomenclatureMapper but is inconsistent between eager
loading and non-eager loading scenarios so will lead to subtle future
bugs.

If an indent level is incorrect, this can be resolved via a data fix on
the `goods_nomenclature_tree_node_overrides` table